Revert "add hack for line spacing in get_line_count()"
This reverts commit 47a31e8f521f5e09a2ee1e134e28f08c2d8a84a2.

+++ scene/gui/rich_text_label.h
@@ -474,7 +474,7 @@ public:
 	bool search(const String &p_string, bool p_from_selection = false, bool p_search_previous = false);
 
 	void scroll_to_line(int p_line);
-	float get_line_count() const;
+	int get_line_count() const;
 	int get_paragraph_count() const;
 	int get_visible_line_count() const;
